refactor: consolidate and simplify test suite
Reduced test files from 17 to 7 (59% reduction) by:
- Removing debug-dom.spec.ts and dynamic-section-height-change.spec.ts
- Consolidating 4 bead positioning tests into bead-positioning.spec.ts
- Consolidating 3 configuration tests into configuration.spec.ts
- Consolidating 3 upper bead tests into upper-beads.spec.ts
- Consolidating 2 UI layout tests into ui-layout.spec.ts

This eliminates redundant test logic while maintaining comprehensive coverage.
