--- tests/dynamic-section-height-change.spec.ts
@@ -1,77 +1,0 @@
-import { test, expect } from '@playwright/test'
-import { TEST_CONFIG } from './test-config'
-
-test.describe('Dynamic Section Height - Configuration Change Simulation', () => {
-  test('should demonstrate failure when bead counts change', async () => {
-    // This test simulates what would happen if we changed the bead counts
-    // but the section heights remained hardcoded
-    
-    console.log('=== Current Configuration ===')
-    console.log('UPPER_BEADS_PER_COLUMN: 1')
-    console.log('LOWER_BEADS_PER_COLUMN: 4')
-    console.log('SECTIONS.UPPER_HEIGHT: 40px (hardcoded)')
-    console.log('SECTIONS.LOWER_HEIGHT: 100px (hardcoded)')
-    
-    console.log('\n=== If we change to 2 upper beads and 5 lower beads ===')
-    const NEW_UPPER_BEADS = 2
-    const NEW_LOWER_BEADS = 5
-    const BEAD_HEIGHT = 20
-    
-    const requiredUpperHeight = (NEW_UPPER_BEADS + 1) * BEAD_HEIGHT
-    const requiredLowerHeight = (NEW_LOWER_BEADS + 1) * BEAD_HEIGHT
-    
-    console.log(`Required upper height: ${requiredUpperHeight}px (${NEW_UPPER_BEADS} beads + 1 space) × ${BEAD_HEIGHT}px`)
-    console.log(`Required lower height: ${requiredLowerHeight}px (${NEW_LOWER_BEADS} beads + 1 space) × ${BEAD_HEIGHT}px`)
-    console.log(`Total required height: ${requiredUpperHeight + requiredLowerHeight}px`)
-    
-    // These assertions demonstrate what would fail
-    console.log('\n=== Expected Failures with Hardcoded Heights ===')
-    
-    // If upper height is still 40px but we need 60px for 2 beads
-    if (40 < requiredUpperHeight) {
-      console.log(`❌ Upper section: 40px < ${requiredUpperHeight}px required (${requiredUpperHeight - 40}px SHORT)`)
-    }
-    
-    // If lower height is still 100px but we need 120px for 5 beads
-    if (100 < requiredLowerHeight) {
-      console.log(`❌ Lower section: 100px < ${requiredLowerHeight}px required (${requiredLowerHeight - 100}px SHORT)`)
-    }
-    
-    // This assertion will fail to demonstrate the problem
-    expect({
-      upperSectionProblem: `40px hardcoded < ${requiredUpperHeight}px needed`,
-      lowerSectionProblem: `100px hardcoded < ${requiredLowerHeight}px needed`,
-      totalShortfall: (requiredUpperHeight - 40) + (requiredLowerHeight - 100)
-    }).toEqual({
-      upperSectionProblem: `40px hardcoded < 60px needed`,
-      lowerSectionProblem: `100px hardcoded < 120px needed`,
-      totalShortfall: 40
-    })
-  })
-
-  test('should show correct formula for dynamic heights', async () => {
-    console.log('=== Correct Dynamic Height Formula ===')
-    console.log('SECTIONS: {')
-    console.log('  UPPER_HEIGHT: (SEMPOA_CONFIG.UPPER_BEADS_PER_COLUMN + 1) * SEMPOA_CONFIG.BEAD.HEIGHT,')
-    console.log('  LOWER_HEIGHT: (SEMPOA_CONFIG.LOWER_BEADS_PER_COLUMN + 1) * SEMPOA_CONFIG.BEAD.HEIGHT,')
-    console.log('}')
-    
-    console.log('\n=== Examples ===')
-    const examples = [
-      { upper: 1, lower: 4, upperHeight: 40, lowerHeight: 100, total: 140 },
-      { upper: 2, lower: 5, upperHeight: 60, lowerHeight: 120, total: 180 },
-      { upper: 3, lower: 6, upperHeight: 80, lowerHeight: 140, total: 220 },
-      { upper: 1, lower: 3, upperHeight: 40, lowerHeight: 80, total: 120 },
-    ]
-    
-    for (const example of examples) {
-      console.log(`${example.upper} upper, ${example.lower} lower beads: ${example.upperHeight}px + ${example.lowerHeight}px = ${example.total}px total`)
-    }
-    
-    // Show that the current hardcoded values only work for one specific case
-    console.log('\n=== Current Hardcoded Values Only Work For ===')
-    console.log('1 upper bead: (1 + 1) × 20 = 40px ✓')
-    console.log('4 lower beads: (4 + 1) × 20 = 100px ✓')
-    console.log('But fail for any other bead count configuration!')
-  })
-})
