--- tests/upper-beads-positioning.spec.ts
@@ -1,184 +1,0 @@
-import { test, expect } from '@playwright/test'
-import { TEST_CONFIG } from './test-config'
-
-test.describe('Upper Beads Positioning', () => {
-  test('should properly position multiple upper beads without intersection', async () => {
-    console.log('=== Testing Upper Bead Positioning Logic ===')
-    
-    // Test the positioning formula for multiple upper beads
-    const BEAD_HEIGHT = 20
-    const scenarios = [
-      { beads: 1, sectionHeight: 40 },
-      { beads: 2, sectionHeight: 60 },
-      { beads: 3, sectionHeight: 80 }
-    ]
-    
-    for (const scenario of scenarios) {
-      console.log(`\nTesting ${scenario.beads} upper bead(s):`)
-      console.log(`Section height: ${scenario.sectionHeight}px`)
-      
-      // Calculate positions for each bead
-      const positions = []
-      for (let row = 0; row < scenario.beads; row++) {
-        // Inactive position: row * BEAD_HEIGHT
-        const inactiveTop = 0 + (row * BEAD_HEIGHT)
-        
-        // Active position: UPPER_ACTIVE_TOP + (row * BEAD_HEIGHT)
-        // Where UPPER_ACTIVE_TOP = separator_center - separator_height/2 - bead_height
-        const separatorCenter = scenario.sectionHeight // Separator at section boundary
-        const separatorHeight = 10
-        const upperActiveTop = separatorCenter - (separatorHeight / 2) - BEAD_HEIGHT
-        const activeTop = upperActiveTop + (row * BEAD_HEIGHT)
-        
-        positions.push({
-          row,
-          inactiveTop,
-          activeTop,
-          inactiveBottom: inactiveTop + BEAD_HEIGHT,
-          activeBottom: activeTop + BEAD_HEIGHT
-        })
-        
-        console.log(`  Bead ${row + 1}: inactive=${inactiveTop}-${inactiveTop + BEAD_HEIGHT}px, active=${activeTop}-${activeTop + BEAD_HEIGHT}px`)
-      }
-      
-      // Check for intersections in inactive state
-      let hasInactiveIntersection = false
-      for (let i = 0; i < positions.length - 1; i++) {
-        const gap = positions[i + 1].inactiveTop - positions[i].inactiveBottom
-        if (gap < 0) {
-          hasInactiveIntersection = true
-          console.log(`  ❌ Inactive intersection: Bead ${i + 1} and ${i + 2} overlap by ${-gap}px`)
-        } else {
-          console.log(`  ✓ Inactive gap: ${gap}px between Bead ${i + 1} and ${i + 2}`)
-        }
-      }
-      
-      // Check for intersections in active state
-      let hasActiveIntersection = false
-      for (let i = 0; i < positions.length - 1; i++) {
-        const gap = positions[i + 1].activeTop - positions[i].activeBottom
-        if (gap < 0) {
-          hasActiveIntersection = true
-          console.log(`  ❌ Active intersection: Bead ${i + 1} and ${i + 2} overlap by ${-gap}px`)
-        } else {
-          console.log(`  ✓ Active gap: ${gap}px between Bead ${i + 1} and ${i + 2}`)
-        }
-      }
-      
-      // Check if all beads fit in section
-      const maxInactiveBottom = Math.max(...positions.map(p => p.inactiveBottom))
-      const maxActiveBottom = Math.max(...positions.map(p => p.activeBottom))
-      
-      const inactiveFitsInSection = maxInactiveBottom <= scenario.sectionHeight
-      const activeFitsInSection = maxActiveBottom <= scenario.sectionHeight
-      
-      console.log(`  ${inactiveFitsInSection ? '✓' : '❌'} Inactive beads fit: max bottom ${maxInactiveBottom}px <= section ${scenario.sectionHeight}px`)
-      console.log(`  ${activeFitsInSection ? '✓' : '❌'} Active beads fit: max bottom ${maxActiveBottom}px <= section ${scenario.sectionHeight}px`)
-      
-      // Assertions
-      expect(hasInactiveIntersection).toBe(false)
-      expect(hasActiveIntersection).toBe(false)
-      expect(inactiveFitsInSection).toBe(true)
-      // Note: Active state may not fit for multiple beads (design consideration)
-      // expect(activeFitsInSection).toBe(true)
-      
-      console.log(`  ✓ All tests passed for ${scenario.beads} bead(s)`)
-    }
-  })
-
-  test('should verify component positioning formula matches expected logic', async ({ page }) => {
-    console.log('=== Testing Component Implementation ===')
-    
-    await page.goto(TEST_CONFIG.BASE_URL)
-    await page.click('button:has-text("Reset")')
-    await page.waitForTimeout(500)
-    
-    // Get current configuration
-    const { SEMPOA_CONFIG } = await import('../src/config/sempoaConfig')
-    
-    console.log(`Current configuration: ${SEMPOA_CONFIG.UPPER_BEADS_PER_COLUMN} upper bead(s) per column`)
-    console.log(`Upper section height: ${SEMPOA_CONFIG.SECTIONS.UPPER_HEIGHT}px`)
-    console.log(`Bead height: ${SEMPOA_CONFIG.BEAD.HEIGHT}px`)
-    
-    // Test first column upper beads
-    const firstColumnBeads = page.locator('.upper-section').first().locator('.absolute')
-    const beadCount = await firstColumnBeads.count()
-    
-    console.log(`First column has ${beadCount} upper bead(s)`)
-    expect(beadCount).toBe(SEMPOA_CONFIG.UPPER_BEADS_PER_COLUMN)
-    
-    // Get positions of all upper beads in first column
-    for (let i = 0; i < beadCount; i++) {
-      const bead = firstColumnBeads.nth(i)
-      const position = await bead.evaluate(el => 
-        parseInt(window.getComputedStyle(el).getPropertyValue('top'))
-      )
-      
-      const expectedPosition = SEMPOA_CONFIG.POSITIONING.UPPER_INACTIVE_TOP + (i * SEMPOA_CONFIG.BEAD.HEIGHT)
-      console.log(`Bead ${i + 1}: actual=${position}px, expected=${expectedPosition}px`)
-      
-      expect(position).toBe(expectedPosition)
-    }
-    
-    console.log('✓ All upper bead positions match expected formula')
-    
-    // If there are multiple beads, test for intersections
-    if (beadCount > 1) {
-      console.log('\n=== Testing Multiple Bead Intersection ===')
-      
-      for (let i = 0; i < beadCount - 1; i++) {
-        const currentBead = firstColumnBeads.nth(i)
-        const nextBead = firstColumnBeads.nth(i + 1)
-        
-        const currentBox = await currentBead.boundingBox()
-        const nextBox = await nextBead.boundingBox()
-        
-        if (currentBox && nextBox) {
-          const gap = nextBox.y - (currentBox.y + currentBox.height)
-          console.log(`Gap between bead ${i + 1} and ${i + 2}: ${gap}px`)
-          
-          expect(gap).toBeGreaterThanOrEqual(0) // No intersection
-        }
-      }
-      
-      console.log('✓ No intersections detected between upper beads')
-    }
-  })
-
-  test('should document behavior when UPPER_BEADS_PER_COLUMN changes', async () => {
-    console.log('=== Documenting Configuration Impact ===')
-    
-    const BEAD_HEIGHT = 20
-    const examples = [
-      { beads: 1, desc: 'Traditional abacus (current)' },
-      { beads: 2, desc: 'Extended configuration' },
-      { beads: 3, desc: 'Large abacus' }
-    ]
-    
-    for (const example of examples) {
-      const sectionHeight = (example.beads + 1) * BEAD_HEIGHT
-      const separatorPosition = sectionHeight
-      const containerHeight = sectionHeight + 100 // + lower section (example)
-      
-      console.log(`\n${example.desc}: ${example.beads} upper bead(s)`)
-      console.log(`  Upper section height: ${sectionHeight}px`)
-      console.log(`  Separator position: ${separatorPosition}px`)
-      console.log(`  Total container height: ${containerHeight}px`)
-      
-      // Show bead positions
-      console.log(`  Bead positions (inactive):`)
-      for (let i = 0; i < example.beads; i++) {
-        const position = i * BEAD_HEIGHT
-        console.log(`    Bead ${i + 1}: ${position}px - ${position + BEAD_HEIGHT}px`)
-      }
-      
-      // Verify calculation
-      const expectedHeight = (example.beads + 1) * BEAD_HEIGHT
-      expect(sectionHeight).toBe(expectedHeight)
-    }
-    
-    console.log('\n✓ All calculations verified')
-    console.log('✓ To change bead count: modify UPPER_BEADS_PER_COLUMN in sempoaConfig.ts')
-    console.log('✓ Section heights and positioning will adjust automatically')
-  })
-})
